% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-endpoint.R, R/class-service.R, R/class-station.R, R/class-status-interval.R, R/class-timeseries.R, R/class-tvp.R, R/generic-methods.R, R/virtual-class-api-resource.R, R/virtual-class-domain-resource.R
\docType{methods}
\name{label,Endpoint-method}
\alias{accessor-methods}
\alias{bbox,Station-method}
\alias{color}
\alias{color,StatusInterval-method}
\alias{color<-}
\alias{color<-,StatusInterval,character_or_NULL-method}
\alias{coordinates,Station-method}
\alias{domainId}
\alias{domainId,DomainResource-method}
\alias{domainId<-}
\alias{domainId<-,DomainResource,character_or_NULL-method}
\alias{endpoint}
\alias{endpoint,ApiResource-method}
\alias{endpoint<-}
\alias{endpoint<-,ApiResource,Endpoint_or_NULL-method}
\alias{firstValue}
\alias{firstValue,Timeseries-method}
\alias{firstValue<-}
\alias{firstValue<-,Timeseries,TVP_or_NULL-method}
\alias{geometry,Station-method}
\alias{geometry<-}
\alias{geometry<-,Station,SpatialPoints_or_NULL-method}
\alias{id}
\alias{id,ApiResource-method}
\alias{label}
\alias{label,ApiResource-method}
\alias{label,Endpoint-method}
\alias{label<-}
\alias{label<-,ApiResource,character_or_NULL-method}
\alias{label<-,Endpoint,character_or_NULL-method}
\alias{lastValue}
\alias{lastValue,Timeseries-method}
\alias{lastValue<-}
\alias{lastValue<-,Timeseries,TVP_or_NULL-method}
\alias{lower}
\alias{lower,StatusInterval-method}
\alias{lower<-}
\alias{lower<-,StatusInterval,numeric_or_NULL-method}
\alias{name}
\alias{name,StatusInterval-method}
\alias{name<-}
\alias{names,ApiResource-method}
\alias{names,Endpoint-method}
\alias{names<-,ApiResource,character_or_NULL-method}
\alias{names<-,Endpoint,character_or_NULL-method}
\alias{quantities}
\alias{quantities,Service-method}
\alias{quantities<-}
\alias{quantities<-,Service,data.frame_or_NULL-method}
\alias{referenceValues}
\alias{referenceValues,Timeseries-method}
\alias{referenceValues<-}
\alias{referenceValues<-,Timeseries,ReferenceValue_or_NULL-method}
\alias{service,DomainResource-method}
\alias{service<-,DomainResource,Service_or_NULL-method}
\alias{serviceURL}
\alias{serviceURL,Service-method}
\alias{serviceURL<-}
\alias{serviceURL<-,Service,character_or_NULL-method}
\alias{statusIntervals}
\alias{statusIntervals,Timeseries-method}
\alias{supportsFirstLatest}
\alias{supportsFirstLatest,Service-method}
\alias{supportsFirstLatest<-}
\alias{supportsFirstLatest<-,Service,logical_or_NULL-method}
\alias{time}
\alias{time,TVP-method}
\alias{type}
\alias{type,Service-method}
\alias{type<-}
\alias{type<-,Service,character_or_NULL-method}
\alias{uom}
\alias{uom,Timeseries-method}
\alias{uom<-}
\alias{uom<-,Timeseries,character_or_NULL-method}
\alias{upper}
\alias{upper,StatusInterval-method}
\alias{upper<-}
\alias{upper<-,StatusInterval,numeric_or_NULL-method}
\alias{value}
\alias{value,TVP-method}
\alias{version}
\alias{version,Service-method}
\alias{version<-}
\alias{version<-,Service,character_or_NULL-method}
\title{Accessor Methods}
\usage{
\S4method{label}{Endpoint}(x)

\S4method{label}{Endpoint,character_or_NULL}(x) <- value

\S4method{names}{Endpoint}(x)

\S4method{names}{Endpoint,character_or_NULL}(x) <- value

\S4method{serviceURL}{Service}(x)

\S4method{version}{Service}(x)

\S4method{type}{Service}(x)

\S4method{supportsFirstLatest}{Service}(x)

\S4method{quantities}{Service}(x)

\S4method{serviceURL}{Service,character_or_NULL}(x) <- value

\S4method{version}{Service,character_or_NULL}(x) <- value

\S4method{type}{Service,character_or_NULL}(x) <- value

\S4method{supportsFirstLatest}{Service,logical_or_NULL}(x) <- value

\S4method{quantities}{Service,data.frame_or_NULL}(x) <- value

\S4method{geometry}{Station}(obj)

\S4method{geometry}{Station,SpatialPoints_or_NULL}(x) <- value

\S4method{coordinates}{Station}(obj, ...)

\S4method{bbox}{Station}(obj)

\S4method{lower}{StatusInterval}(x)

\S4method{lower}{StatusInterval,numeric_or_NULL}(x) <- value

\S4method{upper}{StatusInterval}(x)

\S4method{upper}{StatusInterval,numeric_or_NULL}(x) <- value

\S4method{color}{StatusInterval}(x)

\S4method{color}{StatusInterval,character_or_NULL}(x) <- value

\S4method{name}{StatusInterval}(x)

\S4method{referenceValues}{Timeseries}(x)

\S4method{referenceValues}{Timeseries,ReferenceValue_or_NULL}(x) <- value

\S4method{uom}{Timeseries}(x)

\S4method{uom}{Timeseries,character_or_NULL}(x) <- value

\S4method{firstValue}{Timeseries}(x)

\S4method{firstValue}{Timeseries,TVP_or_NULL}(x) <- value

\S4method{lastValue}{Timeseries}(x)

\S4method{lastValue}{Timeseries,TVP_or_NULL}(x) <- value

\S4method{statusIntervals}{Timeseries}(x)

\S4method{value}{TVP}(x)

\S4method{time}{TVP}(x)

id(x)

label(x)

label(x) <- value

endpoint(x)

endpoint(x) <- value

serviceURL(x)

serviceURL(x) <- value

version(x)

version(x) <- value

type(x)

type(x) <- value

supportsFirstLatest(x)

supportsFirstLatest(x) <- value

quantities(x)

quantities(x) <- value

domainId(x)

domainId(x) <- value

color(x)

color(x) <- value

upper(x)

upper(x) <- value

lower(x)

lower(x) <- value

name(x)

name(x) <- value

time(x, ...)

value(x)

uom(x)

uom(x) <- value

firstValue(x)

firstValue(x) <- value

lastValue(x)

lastValue(x) <- value

statusIntervals(x)

referenceValues(x)

referenceValues(x) <- value

geometry(x) <- value

\S4method{id}{ApiResource}(x)

\S4method{label}{ApiResource}(x)

\S4method{label}{ApiResource,character_or_NULL}(x) <- value

\S4method{names}{ApiResource}(x)

\S4method{names}{ApiResource,character_or_NULL}(x) <- value

\S4method{endpoint}{ApiResource}(x)

\S4method{endpoint}{ApiResource,Endpoint_or_NULL}(x) <- value

\S4method{domainId}{DomainResource}(x)

\S4method{domainId}{DomainResource,character_or_NULL}(x) <- value

\S4method{service}{DomainResource}(x)

\S4method{service}{DomainResource,Service_or_NULL}(x) <- value
}
\description{
Getter and setters for slots.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

