% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-category.R, R/class-timeseries.R, R/generic-methods.R
\docType{methods}
\name{service,Category-method}
\alias{api-relations}
\alias{category}
\alias{category,Timeseries-method}
\alias{category<-}
\alias{category<-,Timeseries,Category_or_NULL-method}
\alias{feature}
\alias{feature,Timeseries-method}
\alias{feature<-}
\alias{feature<-,Timeseries,Feature_or_NULL-method}
\alias{offering}
\alias{offering,Timeseries-method}
\alias{offering<-}
\alias{offering<-,Timeseries,Offering_or_NULL-method}
\alias{phenomenon}
\alias{phenomenon,Timeseries-method}
\alias{phenomenon<-}
\alias{phenomenon<-,Timeseries,Phenomenon_or_NULL-method}
\alias{procedure}
\alias{procedure,Timeseries-method}
\alias{procedure<-}
\alias{procedure<-,Timeseries,Procedure_or_NULL-method}
\alias{service}
\alias{service,Category-method}
\alias{service,Timeseries-method}
\alias{service<-}
\alias{service<-,Category,Service_or_NULL-method}
\alias{service<-,Timeseries,Service_or_NULL-method}
\alias{station}
\alias{station,Timeseries-method}
\alias{station<-}
\alias{station<-,Timeseries,Station_or_NULL-method}
\title{API Relation Getters and Setters}
\usage{
\S4method{service}{Category}(x)

\S4method{service}{Category,Service_or_NULL}(x) <- value

\S4method{service}{Timeseries}(x)

\S4method{service}{Timeseries,Service_or_NULL}(x) <- value

\S4method{feature}{Timeseries}(x)

\S4method{feature}{Timeseries,Feature_or_NULL}(x) <- value

\S4method{offering}{Timeseries}(x)

\S4method{offering}{Timeseries,Offering_or_NULL}(x) <- value

\S4method{category}{Timeseries}(x)

\S4method{category}{Timeseries,Category_or_NULL}(x) <- value

\S4method{procedure}{Timeseries}(x)

\S4method{procedure}{Timeseries,Procedure_or_NULL}(x) <- value

\S4method{station}{Timeseries}(x)

\S4method{station}{Timeseries,Station_or_NULL}(x) <- value

\S4method{phenomenon}{Timeseries}(x)

\S4method{phenomenon}{Timeseries,Phenomenon_or_NULL}(x) <- value

service(x)

service(x) <- value

feature(x)

feature(x) <- value

offering(x)

offering(x) <- value

procedure(x)

procedure(x) <- value

station(x)

station(x) <- value

category(x)

category(x) <- value

phenomenon(x)

phenomenon(x) <- value
}
\arguments{
\item{x}{The \linkS4class{Category}.}
}
\description{
#' Common getter and setter for relations between API resources.
}
\details{
\code{service} gets the \linkS4class{Service} of the resource.

\code{procedure<-} sets the \linkS4class{Procedure} of the resource.

\code{feature} gets the \linkS4class{Feature} of the resource.

\code{feature<-} sets the \linkS4class{Feature} of the resource.

\code{offering} gets the \linkS4class{Offering} of the resource.

\code{offering<-} sets the \linkS4class{Offering} of the resource.

\code{procedure} gets the \linkS4class{Procedure} of the resource.

\code{procedure<-} sets the \linkS4class{Procedure} of the resource.

\code{station} gets the \linkS4class{Station} of the resource.

\code{station<-} sets the \linkS4class{Station} of the resource.

\code{category} gets the \linkS4class{Category} of the resource.

\code{category<-} sets the \linkS4class{Category} of the resource.

\code{phenomenon} gets the \linkS4class{Phenomenon} of the resource.

\code{phenomenon<-} sets the \linkS4class{Phenomenon} of the resource.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

