% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/virtual-class-api-resource.R
\docType{class}
\name{ApiResource-class}
\alias{ApiResource-class}
\alias{as.ApiResource}
\alias{as.list.ApiResource}
\alias{fromURI}
\alias{is.ApiResource}
\alias{unique.ApiResource}
\title{API Resource}
\usage{
is.ApiResource(x)

as.ApiResource(x)

as.list.ApiResource(x, ...)

unique.ApiResource(x, ...)

fromURI(uri)
}
\description{
A virtual class representing a Timeseries API resource.
}
\section{Functions}{
\itemize{
\item \code{is.ApiResource}: Checks whether \code{x} is an \code{ApiResource}.

\item \code{as.ApiResource}: Coerces \code{x} to an \code{ApiResource}.

\item \code{as.list.ApiResource}: Corces the \code{ApiResource} \code{x} into a \code{list}.

\item \code{unique.ApiResource}: Returns the unique \code{ApiResource}s in \code{x}.

\item \code{fromURI}: Parses the given URIs into their respective \code{ApiResource}.
}}
\section{Slots}{

\describe{
\item{\code{id}}{The identifier.}

\item{\code{label}}{The human-readable name.}

\item{\code{endpoint}}{The associated \linkS4class{Endpoint}.}
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

