% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query-methods.R
\name{as.query}
\alias{as.query}
\title{Query Helper}
\usage{
as.query(service = NULL, category = NULL, phenomenon = NULL, crs = NULL,
  bbox = NULL, near = NULL, timespan = NULL, station = NULL,
  locale = NULL)
}
\arguments{
\item{service}{A \linkS4class{Service} or character identifier to filter with.}

\item{category}{A \linkS4class{Category} or character identifier to filter with.}

\item{phenomenon}{A \linkS4class{Phenomenon} or character identifier to filter with.}

\item{crs}{A CRS identifier.}

\item{bbox}{A bouding box character vector.}

\item{near}{A point character vector.}

\item{timespan}{A \linkS4class{lubridate::interval} or a character vector to filter with.}

\item{station}{A \linkS4class{Station} or character identifier to filter with.}

\item{locale}{A locale character vector}
}
\description{
Transforms the parameters to a query list.
}
\keyword{internal}

