% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache-functions.R
\name{cache}
\alias{cache}
\alias{clear.cache}
\alias{get.cache}
\alias{get.cache.keys}
\alias{get.cache.value}
\alias{set.cache}
\alias{set.cache.value}
\alias{set.cache.values}
\title{Cache}
\usage{
clear.cache()

get.cache()

set.cache(cache)

set.cache.value(key, value)

set.cache.values(...)

get.cache.value(key)

get.cache.keys()
}
\arguments{
\item{cache}{A \code{list} object to be used as the cache.}

\item{key}{The key to be used to insert the \code{value}.}

\item{value}{The value to be inserted under \code{key}.}

\item{...}{Named parameters to be inserted into the cache.}
}
\description{
\code{sensorweb4R} cache functions.
}
\details{
Most requests to the Timeseries are cached for performance reasons.
The parsed JSON responses are saved in a global cache object (with
the name \code{sensorweb4R:::CACHE_NAME}). In this list response
objects are saved by their respective URL.
}
\section{Functions}{
\itemize{
\item \code{clear.cache}: Clears the cache.

\item \code{get.cache}: Gets the cache object.

\item \code{set.cache}: Sets the cache object.

\item \code{set.cache.value}: Sets a cache value.

\item \code{set.cache.values}: Sets multiple cache values.

\item \code{get.cache.value}: Gets a cache value.

\item \code{get.cache.keys}: Gets the current cache keys.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

