% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-category.R
\docType{class}
\name{Category-class}
\alias{Category}
\alias{Category-class}
\alias{as.Category}
\alias{is.Category}
\title{Category}
\usage{
is.Category(x)

as.Category(x)

Category(id = character(), label = NULL, endpoint = NULL,
  service = NULL)
}
\description{
Category
}
\section{Functions}{
\itemize{
\item \code{is.Category}: Checks whether \code{x} is a \code{Category}.

\item \code{as.Category}: Coerces \code{x} into a \code{Category}.

\item \code{Category}: Creates a new \code{Category}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

