% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.frame-methods.R
\docType{methods}
\name{data.frame-methods}
\alias{as.data.frame,Category-method}
\alias{as.data.frame,DomainResource-method}
\alias{as.data.frame,Endpoint-method}
\alias{as.data.frame,HttpResource-method}
\alias{as.data.frame,ReferenceValue-method}
\alias{as.data.frame,Service-method}
\alias{as.data.frame,Station-method}
\alias{as.data.frame,StatusInterval-method}
\alias{as.data.frame,TVP-method}
\alias{as.data.frame,Timeseries-method}
\alias{as.data.frame.ApiResource}
\alias{as.data.frame.Category}
\alias{as.data.frame.DomainResource}
\alias{as.data.frame.Endpoint}
\alias{as.data.frame.HttpResource}
\alias{as.data.frame.ReferenceValue}
\alias{as.data.frame.Service}
\alias{as.data.frame.Station}
\alias{as.data.frame.StatusInterval}
\alias{as.data.frame.TVP}
\alias{as.data.frame.Timeseries}
\alias{data.frame-methods}
\title{data.frame coercion}
\usage{
as.data.frame.ApiResource(x, ...)

as.data.frame.Category(x, ...)

\S4method{as.data.frame}{Category}(x, row.names = NULL, optional = FALSE,
  ...)

as.data.frame.DomainResource(x, ...)

\S4method{as.data.frame}{DomainResource}(x, row.names = NULL,
  optional = FALSE, ...)

as.data.frame.Endpoint(x, ...)

\S4method{as.data.frame}{Endpoint}(x, row.names = NULL, optional = FALSE,
  ...)

as.data.frame.HttpResource(x, ...)

\S4method{as.data.frame}{HttpResource}(x, row.names = NULL,
  optional = FALSE, ...)

as.data.frame.ReferenceValue(x, ...)

\S4method{as.data.frame}{ReferenceValue}(x, row.names = NULL,
  optional = FALSE, ...)

as.data.frame.Service(x, ...)

\S4method{as.data.frame}{Service}(x, row.names = NULL, optional = FALSE,
  ...)

as.data.frame.Station(x, ...)

\S4method{as.data.frame}{Station}(x, row.names = NULL, optional = FALSE,
  ...)

as.data.frame.StatusInterval(x, ...)

\S4method{as.data.frame}{StatusInterval}(x, row.names = NULL,
  optional = FALSE, ...)

as.data.frame.Timeseries(x, ...)

\S4method{as.data.frame}{Timeseries}(x, row.names = NULL, optional = FALSE,
  ...)

as.data.frame.TVP(x, ...)

\S4method{as.data.frame}{TVP}(x, row.names = NULL, optional = FALSE, ...)
}
\description{
Methods to coerce to a  \code{data.frame}.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

