% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-matrix.R
\docType{methods}
\name{distance-matrix-methods}
\alias{distance}
\alias{distance,DistanceStation-method}
\alias{distance-matrix-methods}
\alias{distanceMatrix}
\alias{distanceMatrix,SpatialPoints-method}
\alias{distanceMatrix,Station-method}
\alias{nearestStations}
\alias{nearestStations,Station-method}
\alias{readDistanceMatrix}
\alias{saveDistanceMatrix}
\title{Distance matrix methods}
\usage{
saveDistanceMatrix(dm, file)

readDistanceMatrix(file)

nearestStations(x, ...)

distanceMatrix(x, ...)

distance(x)

\S4method{nearestStations}{Station}(x, stations = NULL, dm = NULL, n = 1,
  ...)

\S4method{distance}{DistanceStation}(x)

\S4method{distanceMatrix}{SpatialPoints}(x, ...)

\S4method{distanceMatrix}{Station}(x, ...)
}
\arguments{
\item{dm}{the distance matrix}

\item{file}{the file name}

\item{x}{the object to find the nearest stations for}

\item{stations}{the \linkS4class{Station}s for the distance
calculation (requested if \code{NULL})}

\item{n}{the number of stations that should be returned}

\item{file}{the file name}

\item{x}{The object to create a distance matrix for.}

\item{dm}{the distance matrix for the stations (generated
if \code{NULL})}
}
\value{
a subset of \code{stations} orderd by their distance
         to \code{x}, enriched with an additional
         \code{distance} attribute

the distance matrix of class \code{dist}
}
\description{
Methods to calculate distances between \linkS4class{Station}s
and to find the nearest \code{Station}s.

\code{saveDistanceMatrix} a distance matrix to disk.

\code{readDistanceMatrix} reads the a distance matrix from disk.

\code{nearestStations} gets the \code{n} stations nearest
to \code{station}.

\code{distanceMatrix} creates a distance matrix for the supplied object.
}
\examples{
\dontrun{
saveDistanceMatrix(dm, 'distance_matrix.rds')
}
\dontrun{
dm <- readDistanceMatrix('distance_matrix.rds')
}
}

