% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/virtual-class-domain-resource.R
\docType{class}
\name{DomainResource-class}
\alias{DomainResource}
\alias{DomainResource-class}
\alias{as.DomainResource}
\alias{is.DomainResource}
\title{Domain Resource}
\usage{
DomainResource(type, id = character(), label = NULL, domainId = NULL,
  service = NULL, endpoint = NULL)

is.DomainResource(x)

as.DomainResource(x)
}
\description{
Domain Resource
}
\section{Functions}{
\itemize{
\item \code{DomainResource}: Constructs a new \code{DomainResource}.

\item \code{is.DomainResource}: Checks if \code{x} is a \code{DomainResource}.

\item \code{as.DomainResource}: Coerces \code{x} to a \code{DomainResource}.
}}

