% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-endpoint.R
\docType{class}
\name{Endpoint-class}
\alias{Endpoint}
\alias{Endpoint-class}
\alias{as.Endpoint}
\alias{is.Endpoint}
\title{Endpoint class}
\usage{
is.Endpoint(x)

as.Endpoint(x)

Endpoint(url = character(), label = NULL, ...)
}
\description{
A class representing a Timeseries API endpoint.
}
\section{Functions}{
\itemize{
\item \code{is.Endpoint}: Checks wether \code{x} is an \code{Endpoint}.

\item \code{as.Endpoint}: Coerces \code{x} into an \code{Endpoint}.

\item \code{Endpoint}: Constructs a new \code{Endpoint}.
}}
\section{Slots}{

\describe{
\item{\code{url}}{The URL.}

\item{\code{label}}{A human readable name.}
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

