% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-feature.R
\docType{class}
\name{Feature-class}
\alias{Feature}
\alias{Feature-class}
\alias{as.Feature}
\alias{is.Feature}
\title{Feature}
\usage{
is.Feature(x)

as.Feature(x)

Feature(...)
}
\description{
Represents a feature.
}
\section{Functions}{
\itemize{
\item \code{is.Feature}: Checks whether \code{x} is a \code{Feature}.

\item \code{as.Feature}: Coerces \code{x} into a \code{Feature}.

\item \code{Feature}: Constructs a new \code{Feature}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

