% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic-methods.R, R/query-methods.R
\docType{methods}
\name{fetch-methods}
\alias{fetch}
\alias{fetch,Category-method}
\alias{fetch,DomainResource-method}
\alias{fetch,Service-method}
\alias{fetch,Station-method}
\alias{fetch,Timeseries-method}
\alias{fetch-methods}
\title{Fetch methods}
\usage{
fetch(x, ...)

\S4method{fetch}{Service}(x, ...)

\S4method{fetch}{DomainResource}(x, ...)

\S4method{fetch}{Category}(x, ...)

\S4method{fetch}{Station}(x, ...)

\S4method{fetch}{Timeseries}(x, ...)
}
\arguments{
\item{x}{the API resource to fetch}

\item{...}{Further arguments that may be needed by the actual implementation}
}
\value{
an object of the same class as \code{x} with all meta data added
}
\description{
\code{fetch} fetches recursively all metadata for the supplied API resource.
}
\examples{
\dontrun{
fetch(Timeseries(id = ids, endpoint = Endpoint(e)))
}
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

