% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stations.R
\name{get_nearest_stations}
\alias{get_nearest_stations}
\title{Get Nearest Stations}
\usage{
get_nearest_stations(station, stations, dm = NULL, n = 1)
}
\arguments{
\item{station}{the station identifier}

\item{stations}{the \code{SpatialPointsDataFrame} containing
the stations}

\item{dm}{the distance matrix for the stations (generated
if \code{NULL})}

\item{n}{the number of stations that should be returned}
}
\value{
a subset of \code{stations} orderd by their distance
         to \code{station}, enriched with an additional
         \code{dist} attribute
}
\description{
\code{get_nearest_stations} gets the \code{n} stations nearest
to \code{station}.
}
\examples{
\dontrun{
url <- 'http://sensorweb.demo.52north.org/sensorwebclient-webapp-stable'
stations <- get_stations(url, service = 'srv_6d9ccea8d609ecb74d4a512922bb7cee')
dm <- create_distance_matrix(stations)
get_nearest_stations(station, stations, dm, n=5)
}
}

