% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has_license_header}
\alias{has_license_header}
\title{License header check}
\usage{
has_license_header(pkg = ".", header = "inst/license-header")
}
\arguments{
\item{pkg}{package name, used in \code{devtools:::find_code(..)}.}

\item{header}{a path to the header file to be used}
}
\value{
\code{TRUE} if all files have the header, \code{FALSE} otherwise.
}
\description{
\code{has_license_header} checks if license header is present in all the R source code files of the package using \code{licesen_header()}.
}
\examples{
\dontrun{
 has_license_header()
}
}

