% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/virtual-class-http-resource.R
\docType{class}
\name{HttpResource-class}
\alias{HttpResource-class}
\alias{as.HttpResource}
\alias{is.HttpResource}
\title{HTTP Resource}
\usage{
is.HttpResource(x)

as.HttpResource(x)
}
\description{
Virtual class representing a HTTP resource.
}
\section{Functions}{
\itemize{
\item \code{is.HttpResource}: Checks whether \code{x} is a \code{HttpResource}.

\item \code{as.HttpResource}: Coerces \code{x} into a \code{HttpResource}.
}}

