% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-endpoint.R, R/class-status-interval.R, R/class-tvp.R, R/generic-methods.R, R/virtual-class-api-resource.R
\docType{methods}
\name{length,Endpoint-method}
\alias{length,ApiResource-method}
\alias{length,Endpoint-method}
\alias{length,StatusInterval-method}
\alias{length,TVP-method}
\alias{length-methods}
\title{length Methods}
\usage{
\S4method{length}{Endpoint}(x)

\S4method{length}{StatusInterval}(x)

\S4method{length}{TVP}(x)

\S4method{length}{ApiResource}(x)
}
\description{
Gets the length/count of an resource.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

