% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-offering.R
\docType{class}
\name{Offering-class}
\alias{Offering}
\alias{Offering-class}
\alias{as.Offering}
\alias{is.Offering}
\title{Offering}
\usage{
is.Offering(x)

as.Offering(x)

Offering(...)
}
\description{
Represents a offering.
}
\section{Functions}{
\itemize{
\item \code{is.Offering}: Checks whether \code{x} is a \code{Offering}.

\item \code{as.Offering}: Coerces \code{x} into a \code{Offering}.

\item \code{Offering}: Constructs a new \code{Offering}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

