% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-phenomenon.R
\docType{class}
\name{Phenomenon-class}
\alias{Phenomenon}
\alias{Phenomenon-class}
\alias{as.Phenomenon}
\alias{is.Phenomenon}
\title{Phenomenon}
\usage{
is.Phenomenon(x)

as.Phenomenon(x)

Phenomenon(...)
}
\description{
Represents a phenomenon.
}
\section{Functions}{
\itemize{
\item \code{is.Phenomenon}: Checks whether \code{x} is a \code{Phenomenon}.

\item \code{as.Phenomenon}: Coerces \code{x} into a \code{Phenomenon}.

\item \code{Phenomenon}: Constructs a new \code{Phenomenon}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Procedure}},
  \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

