% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-procedure.R
\docType{class}
\name{Procedure-class}
\alias{Procedure}
\alias{Procedure-class}
\alias{as.Procedure}
\alias{is.Procedure}
\title{Procedure}
\usage{
is.Procedure(x)

as.Procedure(x)

Procedure(...)
}
\description{
Represents a procedure.
}
\section{Functions}{
\itemize{
\item \code{is.Procedure}: Checks whether \code{x} is a \code{Procedure}.

\item \code{as.Procedure}: Coerces \code{x} into a \code{Procedure}.

\item \code{Procedure}: Constructs a new \code{Procedure}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

