% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic-methods.R, R/query-methods.R
\docType{methods}
\name{query-methods}
\alias{categories}
\alias{categories,Endpoint-method}
\alias{categories,Service-method}
\alias{categories,Station-method}
\alias{features}
\alias{features,Endpoint-method}
\alias{features,Service-method}
\alias{features,Station-method}
\alias{getData}
\alias{getData,Timeseries-method}
\alias{offerings}
\alias{offerings,Endpoint-method}
\alias{offerings,Service-method}
\alias{offerings,Station-method}
\alias{phenomena}
\alias{phenomena,Endpoint-method}
\alias{phenomena,Service-method}
\alias{phenomena,Station-method}
\alias{procedures}
\alias{procedures,Endpoint-method}
\alias{procedures,Service-method}
\alias{procedures,Station-method}
\alias{query-methods}
\alias{services}
\alias{services,Endpoint-method}
\alias{services,Station-method}
\alias{stations}
\alias{stations,Endpoint-method}
\alias{stations,Service-method}
\alias{timeseries}
\alias{timeseries,Endpoint-method}
\alias{timeseries,Service-method}
\alias{timeseries,Station-method}
\title{Query Methods}
\usage{
stations(x, ...)

timeseries(x, ...)

services(x, ...)

offerings(x, ...)

procedures(x, ...)

phenomena(x, ...)

categories(x, ...)

features(x, ...)

getData(x, ...)

\S4method{stations}{Endpoint}(x, ...)

\S4method{phenomena}{Endpoint}(x, ...)

\S4method{categories}{Endpoint}(x, ...)

\S4method{features}{Endpoint}(x, ...)

\S4method{offerings}{Endpoint}(x, ...)

\S4method{procedures}{Endpoint}(x, ...)

\S4method{timeseries}{Endpoint}(x, ...)

\S4method{services}{Endpoint}(x, ...)

\S4method{stations}{Service}(x, ...)

\S4method{phenomena}{Service}(x, ...)

\S4method{features}{Service}(x, ...)

\S4method{categories}{Service}(x, ...)

\S4method{timeseries}{Service}(x, ...)

\S4method{offerings}{Service}(x, ...)

\S4method{procedures}{Service}(x, ...)

\S4method{services}{Station}(x, ...)

\S4method{phenomena}{Station}(x, ...)

\S4method{features}{Station}(x, ...)

\S4method{categories}{Station}(x, ...)

\S4method{timeseries}{Station}(x, ...)

\S4method{offerings}{Station}(x, ...)

\S4method{procedures}{Station}(x, ...)

\S4method{getData}{Timeseries}(x, generalize = FALSE, timespan = NULL, ...)
}
\arguments{
\item{timespan}{A \linkS4class{lubridate::interval} or a character vector to filter with.}
}
\description{
Queries an \linkS4class{Endpoint}.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

