% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-category.R, R/class-endpoint.R, R/class-reference-value.R, R/class-service.R, R/class-station.R, R/class-status-interval.R, R/class-timeseries.R, R/class-tvp.R, R/generic-methods.R, R/virtual-class-domain-resource.R
\docType{methods}
\name{rbind2,Category,Category-method}
\alias{rbind2,ANY,ANY-method}
\alias{rbind2,ANY,Category-method}
\alias{rbind2,ANY,DomainResource-method}
\alias{rbind2,ANY,Endpoint-method}
\alias{rbind2,ANY,ReferenceValue-method}
\alias{rbind2,ANY,Service-method}
\alias{rbind2,ANY,SpatialPoints-method}
\alias{rbind2,ANY,Station-method}
\alias{rbind2,ANY,StatusInterval-method}
\alias{rbind2,ANY,TVP-method}
\alias{rbind2,ANY,Timeseries-method}
\alias{rbind2,Category,ANY-method}
\alias{rbind2,Category,Category-method}
\alias{rbind2,DomainResource,ANY-method}
\alias{rbind2,DomainResource,DomainResource-method}
\alias{rbind2,Endpoint,ANY-method}
\alias{rbind2,Endpoint,Endpoint-method}
\alias{rbind2,ReferenceValue,ANY-method}
\alias{rbind2,ReferenceValue,ReferenceValue-method}
\alias{rbind2,Service,ANY-method}
\alias{rbind2,Service,Service-method}
\alias{rbind2,SpatialPoints,ANY-method}
\alias{rbind2,SpatialPoints,SpatialPoints-method}
\alias{rbind2,Station,ANY-method}
\alias{rbind2,Station,Station-method}
\alias{rbind2,StatusInterval,ANY-method}
\alias{rbind2,StatusInterval,StatusInterval-method}
\alias{rbind2,TVP,ANY-method}
\alias{rbind2,TVP,TVP-method}
\alias{rbind2,Timeseries,ANY-method}
\alias{rbind2,Timeseries,Timeseries-method}
\alias{rbind2-methods}
\title{rbind2 Methods}
\usage{
\S4method{rbind2}{Category,Category}(x, y)

\S4method{rbind2}{Category,ANY}(x, y)

\S4method{rbind2}{ANY,Category}(x, y)

\S4method{rbind2}{Endpoint,Endpoint}(x, y)

\S4method{rbind2}{Endpoint,ANY}(x, y)

\S4method{rbind2}{ANY,Endpoint}(x, y)

\S4method{rbind2}{ReferenceValue,ReferenceValue}(x, y)

\S4method{rbind2}{ReferenceValue,ANY}(x, y)

\S4method{rbind2}{ANY,ReferenceValue}(x, y)

\S4method{rbind2}{Service,Service}(x, y)

\S4method{rbind2}{Service,ANY}(x, y)

\S4method{rbind2}{ANY,Service}(x, y)

\S4method{rbind2}{Station,Station}(x, y)

\S4method{rbind2}{Station,ANY}(x, y)

\S4method{rbind2}{ANY,Station}(x, y)

\S4method{rbind2}{SpatialPoints,SpatialPoints}(x, y)

\S4method{rbind2}{SpatialPoints,ANY}(x, y)

\S4method{rbind2}{ANY,SpatialPoints}(x, y)

\S4method{rbind2}{StatusInterval,StatusInterval}(x, y)

\S4method{rbind2}{StatusInterval,ANY}(x, y)

\S4method{rbind2}{ANY,StatusInterval}(x, y)

\S4method{rbind2}{Timeseries,Timeseries}(x, y)

\S4method{rbind2}{Timeseries,ANY}(x, y)

\S4method{rbind2}{ANY,Timeseries}(x, y)

\S4method{rbind2}{TVP,TVP}(x, y)

\S4method{rbind2}{TVP,ANY}(x, y)

\S4method{rbind2}{ANY,TVP}(x, y)

\S4method{rbind2}{DomainResource,DomainResource}(x, y)

\S4method{rbind2}{DomainResource,ANY}(x, y)

\S4method{rbind2}{ANY,DomainResource}(x, y)

\S4method{rbind2}{ANY,ANY}(x, y)
}
\description{
\code{rbind2} combines to objects into one.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

