% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-reference-value.R
\docType{class}
\name{ReferenceValue-class}
\alias{ReferenceValue}
\alias{ReferenceValue-class}
\alias{as.ReferenceValue}
\alias{is.ReferenceValue}
\title{ReferenceValue}
\usage{
is.ReferenceValue(x)

as.ReferenceValue(x)

ReferenceValue(id = character(), label = NULL, endpoint = NULL,
  time = NULL, value = NULL)
}
\description{
Represents a reference value of a \linkS4class{Timeseries}.
}
\section{Functions}{
\itemize{
\item \code{is.ReferenceValue}: Checks whether \code{x} is a \code{ReferenceValue}.

\item \code{as.ReferenceValue}: Coerces \code{x} into a \code{ReferenceValue}.

\item \code{ReferenceValue}: Constructs a new \code{ReferenceValue}
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

