% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-category.R, R/class-endpoint.R, R/class-reference-value.R, R/class-service.R, R/class-station.R, R/class-status-interval.R, R/class-timeseries.R, R/class-tvp.R, R/generic-methods.R, R/virtual-class-domain-resource.R
\docType{methods}
\name{rep,Category-method}
\alias{rep,Category-method}
\alias{rep,DomainResource-method}
\alias{rep,Endpoint-method}
\alias{rep,ReferenceValue-method}
\alias{rep,Service-method}
\alias{rep,SpatialPoints-method}
\alias{rep,Station-method}
\alias{rep,StatusInterval-method}
\alias{rep,TVP-method}
\alias{rep,Timeseries-method}
\alias{rep-methods}
\title{rep Methods}
\usage{
\S4method{rep}{Category}(x, ...)

\S4method{rep}{Endpoint}(x, ...)

\S4method{rep}{ReferenceValue}(x, ...)

\S4method{rep}{Service}(x, ...)

\S4method{rep}{Station}(x, ...)

\S4method{rep}{SpatialPoints}(x, ...)

\S4method{rep}{SpatialPoints}(x, ...)

\S4method{rep}{StatusInterval}(x, ...)

\S4method{rep}{Timeseries}(x, ...)

\S4method{rep}{TVP}(x, ...)

\S4method{rep}{DomainResource}(x, ...)
}
\description{
S4 generic version of \code{rep}.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

