% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sensorweb4R}
\alias{sensorweb4R}
\alias{sensorweb4R-package}
\title{Sensor web for R.}
\description{
Sensor web for R.
}
\section{Package options}{

sensorweb4R uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{sensorweb4R.defaultApiVersion}: default version of API to use
  \item \code{sensorweb4R.defaultLoggerThreshold}: default logger used based on the \code{futile.logger} package, logs to console
  \item \code{sensorweb4R.loggerLayout}: layout string used in \code{futile.logger::flog.layout(..)}
}
}

\section{Debugging}{

sensorweb4R uses \code{futile.logger} for logging. To increase the log level of this package call \code{flog.threshold("<level>", name = "sensorweb4R")}. For more information about available logging levels or advanced logger configuration, see \code{?futile.logger}.
}

