% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api-endpoints.R
\name{sensorweb_api_endpoints}
\alias{sensorweb_api_endpoints}
\title{Example API endpoints.}
\usage{
sensorweb_api_endpoints()
}
\value{
R object with the further endpoints offered by the service
}
\description{
\code{sensorweb_api_endpoints} returns a list of endpoints that can be used for testing.
}
\examples{
sensorweb_api_endpoints()
sensorweb_index(sensorweb_api_endpoints()[[1]])
}
\author{
Daniel Nuest \email{d.nuest@52north.org}
}

