% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sensorweb_index}
\alias{sensorweb_index}
\title{Sensorweb API index}
\usage{
sensorweb_index(url, ...)
}
\arguments{
\item{url}{The API endpoint to use}

\item{...}{Further arguments passed on to \code{GET}}
}
\value{
R object with the further endpoints offered by the service
}
\description{
\code{sensorweb_index} returns an object representing the index endpoint of
the sensor web API.
}
\examples{
sensorweb_index(sensorweb_api_endpoints()[[1]])
}
\author{
Daniel Nuest \email{d.nuest@52north.org}
}

