% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SensorwebEndpoint-class}
\alias{SensorwebEndpoint-class}
\title{An S4 class to represent a sensor web API endpoint}
\description{
An S4 class to represent a sensor web API endpoint
}
\section{Slots}{

\describe{
\item{\code{url}}{The URL for the API instance}
}}

