% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-service.R
\docType{class}
\name{Service-class}
\alias{Service}
\alias{Service-class}
\alias{as.Service}
\alias{is.Service}
\title{Service}
\usage{
Service(id = character(), label = NULL, serviceURL = NULL,
  version = NULL, type = NULL, supportsFirstLatest = NULL,
  quantities = NULL, endpoint = NULL)

is.Service(x)

as.Service(x)
}
\description{
Service
}
\section{Functions}{
\itemize{
\item \code{Service}: Constructs a new \code{Service} object.

\item \code{is.Service}: Checks if \code{x} is a \code{Service}.

\item \code{as.Service}: Coerces \code{x} into a \code{Service}.
}}
\section{Slots}{

\describe{
\item{\code{serviceURL}}{the URL of the service}

\item{\code{version}}{the version of the service}

\item{\code{type}}{the type of the service}

\item{\code{supportsFirstLatest}}{wether the service supports first/lastest values}

\item{\code{quantities}}{a \code{data.frame} containing the quantities of the offered resources}
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}};
  \code{\link{Timeseries}}, \code{\link{Timeseries-class}},
  \code{\link{as.Timeseries}}, \code{\link{is.Timeseries}}
}

