% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-status-interval.R
\docType{class}
\name{StatusInterval-class}
\alias{StatusInterval}
\alias{StatusInterval-class}
\alias{as.StatusInterval}
\alias{is.StatusInterval}
\title{StatusInterval}
\usage{
is.StatusInterval(x)

as.StatusInterval(x)

StatusInterval(name = character(), lower = NULL, upper = NULL,
  color = NULL)
}
\description{
StatusInterval
}
\section{Functions}{
\itemize{
\item \code{is.StatusInterval}: Checks whether \code{x} is a \code{StatusInterval}.

\item \code{as.StatusInterval}: Coerces \code{x} into a \code{StatusInterval}.

\item \code{StatusInterval}: Constructs a new \code{StatusInterval}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

