% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic-methods.R, R/string-methods.R
\docType{methods}
\name{string-methods}
\alias{as.character,ApiResource-method}
\alias{as.character,Category-method}
\alias{as.character,DomainResource-method}
\alias{as.character,Endpoint-method}
\alias{as.character,Feature-method}
\alias{as.character,HttpResource-method}
\alias{as.character,Offering-method}
\alias{as.character,Phenomenon-method}
\alias{as.character,Procedure-method}
\alias{as.character,ReferenceValue-method}
\alias{as.character,Service-method}
\alias{as.character,Station-method}
\alias{as.character,StatusInterval-method}
\alias{as.character,TVP-method}
\alias{as.character,Timeseries-method}
\alias{as.character.ApiResource}
\alias{as.character.Category}
\alias{as.character.DomainResource}
\alias{as.character.Endpoint}
\alias{as.character.Feature}
\alias{as.character.HttpResource}
\alias{as.character.Offering}
\alias{as.character.Phenomenon}
\alias{as.character.Procedure}
\alias{as.character.ReferenceValue}
\alias{as.character.Service}
\alias{as.character.Station}
\alias{as.character.StatusInterval}
\alias{as.character.TVP}
\alias{as.character.Timeseries}
\alias{print,ApiResource-method}
\alias{print,Category-method}
\alias{print,DomainResource-method}
\alias{print,Endpoint-method}
\alias{print,Feature-method}
\alias{print,HttpResource-method}
\alias{print,Offering-method}
\alias{print,Phenomenon-method}
\alias{print,Procedure-method}
\alias{print,ReferenceValue-method}
\alias{print,Service-method}
\alias{print,Station-method}
\alias{print,StatusInterval-method}
\alias{print,TVP-method}
\alias{print,Timeseries-method}
\alias{print.ApiResource}
\alias{print.Category}
\alias{print.DomainResource}
\alias{print.Endpoint}
\alias{print.Feature}
\alias{print.HttpResource}
\alias{print.Offering}
\alias{print.Phenomenon}
\alias{print.Procedure}
\alias{print.ReferenceValue}
\alias{print.Service}
\alias{print.Station}
\alias{print.StatusInterval}
\alias{print.TVP}
\alias{print.Timeseries}
\alias{show,ApiResource-method}
\alias{show,Category-method}
\alias{show,DomainResource-method}
\alias{show,Endpoint-method}
\alias{show,Feature-method}
\alias{show,HttpResource-method}
\alias{show,Offering-method}
\alias{show,Phenomenon-method}
\alias{show,Procedure-method}
\alias{show,ReferenceValue-method}
\alias{show,Service-method}
\alias{show,Station-method}
\alias{show,StatusInterval-method}
\alias{show,TVP-method}
\alias{show,Timeseries-method}
\alias{string-methods}
\alias{toString,ApiResource-method}
\alias{toString,Category-method}
\alias{toString,DomainResource-method}
\alias{toString,Endpoint-method}
\alias{toString,Feature-method}
\alias{toString,HttpResource-method}
\alias{toString,Offering-method}
\alias{toString,Phenomenon-method}
\alias{toString,Procedure-method}
\alias{toString,ReferenceValue-method}
\alias{toString,Service-method}
\alias{toString,Station-method}
\alias{toString,StatusInterval-method}
\alias{toString,TVP-method}
\alias{toString,Timeseries-method}
\alias{toString.ApiResource}
\alias{toString.Category}
\alias{toString.DomainResource}
\alias{toString.Endpoint}
\alias{toString.Feature}
\alias{toString.HttpResource}
\alias{toString.Offering}
\alias{toString.Phenomenon}
\alias{toString.Procedure}
\alias{toString.ReferenceValue}
\alias{toString.Service}
\alias{toString.Station}
\alias{toString.StatusInterval}
\alias{toString.TVP}
\alias{toString.Timeseries}
\title{String Methods}
\usage{
\method{as.character}{ApiResource}(x, ...)

\S4method{as.character}{ApiResource}(x, ...)

print.ApiResource(x, ...)

\S4method{print}{ApiResource}(x, ...)

\S4method{show}{ApiResource}(object)

toString.ApiResource(x, ...)

\S4method{toString}{ApiResource}(x, ...)

\method{as.character}{Category}(x, ...)

\S4method{as.character}{Category}(x, ...)

print.Category(x, ...)

\S4method{print}{Category}(x, ...)

\S4method{show}{Category}(object)

toString.Category(x, ...)

\S4method{toString}{Category}(x, ...)

\method{as.character}{DomainResource}(x, ...)

\S4method{as.character}{DomainResource}(x, ...)

print.DomainResource(x, ...)

\S4method{print}{DomainResource}(x, ...)

\S4method{show}{DomainResource}(object)

toString.DomainResource(x, ...)

\S4method{toString}{DomainResource}(x, ...)

\method{as.character}{Endpoint}(x, ...)

\S4method{as.character}{Endpoint}(x, ...)

print.Endpoint(x, ...)

\S4method{print}{Endpoint}(x, ...)

\S4method{show}{Endpoint}(object)

toString.Endpoint(x, ...)

\S4method{toString}{Endpoint}(x, ...)

\method{as.character}{Feature}(x, ...)

\S4method{as.character}{Feature}(x, ...)

print.Feature(x, ...)

\S4method{print}{Feature}(x, ...)

\S4method{show}{Feature}(object)

toString.Feature(x, ...)

\S4method{toString}{Feature}(x, ...)

\method{as.character}{HttpResource}(x, ...)

\S4method{as.character}{HttpResource}(x, ...)

print.HttpResource(x, ...)

\S4method{print}{HttpResource}(x, ...)

\S4method{show}{HttpResource}(object)

toString.HttpResource(x, ...)

\S4method{toString}{HttpResource}(x, ...)

\method{as.character}{Offering}(x, ...)

\S4method{as.character}{Offering}(x, ...)

print.Offering(x, ...)

\S4method{print}{Offering}(x, ...)

\S4method{show}{Offering}(object)

toString.Offering(x, ...)

\S4method{toString}{Offering}(x, ...)

\method{as.character}{Phenomenon}(x, ...)

\S4method{as.character}{Phenomenon}(x, ...)

print.Phenomenon(x, ...)

\S4method{print}{Phenomenon}(x, ...)

\S4method{show}{Phenomenon}(object)

toString.Phenomenon(x, ...)

\S4method{toString}{Phenomenon}(x, ...)

\method{as.character}{Procedure}(x, ...)

\S4method{as.character}{Procedure}(x, ...)

print.Procedure(x, ...)

\S4method{print}{Procedure}(x, ...)

\S4method{show}{Procedure}(object)

toString.Procedure(x, ...)

\S4method{toString}{Procedure}(x, ...)

\method{as.character}{ReferenceValue}(x, ...)

\S4method{as.character}{ReferenceValue}(x, ...)

print.ReferenceValue(x, ...)

\S4method{print}{ReferenceValue}(x, ...)

\S4method{show}{ReferenceValue}(object)

toString.ReferenceValue(x, ...)

\S4method{toString}{ReferenceValue}(x, ...)

\method{as.character}{Service}(x, ...)

\S4method{as.character}{Service}(x, ...)

print.Service(x, ...)

\S4method{print}{Service}(x, ...)

\S4method{show}{Service}(object)

toString.Service(x, ...)

\S4method{toString}{Service}(x, ...)

\method{as.character}{Station}(x, ...)

\S4method{as.character}{Station}(x, ...)

print.Station(x, ...)

\S4method{print}{Station}(x, ...)

\S4method{show}{Station}(object)

toString.Station(x, ...)

\S4method{toString}{Station}(x, ...)

\method{as.character}{StatusInterval}(x, ...)

\S4method{as.character}{StatusInterval}(x, ...)

print.StatusInterval(x, ...)

\S4method{print}{StatusInterval}(x, ...)

\S4method{show}{StatusInterval}(object)

toString.StatusInterval(x, ...)

\S4method{toString}{StatusInterval}(x, ...)

\method{as.character}{Timeseries}(x, ...)

\S4method{as.character}{Timeseries}(x, ...)

print.Timeseries(x, ...)

\S4method{print}{Timeseries}(x, ...)

\S4method{show}{Timeseries}(object)

toString.Timeseries(x, ...)

\S4method{toString}{Timeseries}(x, ...)

\method{as.character}{TVP}(x, ...)

\S4method{as.character}{TVP}(x, ...)

print.TVP(x, ...)

\S4method{print}{TVP}(x, ...)

\S4method{show}{TVP}(object)

toString.TVP(x, ...)

\S4method{toString}{TVP}(x, ...)
}
\description{
Methods and related functions to coerce objects from or to characters.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

