% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract-and-replace-methods.R, R/generic-methods.R
\docType{methods}
\name{[,Category,ANY,missing-method}
\alias{[,Category,ANY,missing-method}
\alias{[,DomainResource,ANY,missing-method}
\alias{[,Endpoint,ANY,missing-method}
\alias{[,ReferenceValue,ANY,missing-method}
\alias{[,Service,ANY,missing-method}
\alias{[,Station,ANY,missing-method}
\alias{[,StatusInterval,ANY,missing-method}
\alias{[,TVP,ANY,missing-method}
\alias{[,Timeseries,ANY,missing-method}
\alias{[<-,Category,ANY,missing,ANY-method}
\alias{[<-,DomainResource,ANY,missing,ANY-method}
\alias{[<-,Endpoint,ANY,missing,ANY-method}
\alias{[<-,ReferenceValue,ANY,missing,ANY-method}
\alias{[<-,Service,ANY,missing,ANY-method}
\alias{[<-,Station,ANY,missing,ANY-method}
\alias{[<-,StatusInterval,ANY,missing,ANY-method}
\alias{[<-,TVP,ANY,missing,ANY-method}
\alias{[<-,Timeseries,ANY,missing,ANY-method}
\alias{subset-methods}
\title{Extraction and Replacement Methods}
\usage{
\S4method{[}{Category,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{Category,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{DomainResource,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{DomainResource,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{Endpoint,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{Endpoint,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{ReferenceValue,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{ReferenceValue,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{Service,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{Service,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{Station,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{Station,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{StatusInterval,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{StatusInterval,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{Timeseries,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{Timeseries,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{TVP,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{TVP,ANY,missing,ANY}(x, i, j) <- value
}
\description{
Methods for extraction and replacement.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

