% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Timeseries-class}
\alias{Timeseries-class}
\title{An S4 class to represent a time series.}
\description{
An S4 class to represent a time series.
}
\section{Slots}{

\describe{
\item{\code{id}}{The identifier for the timeseries}
}}

