% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timeseries}
\alias{timeseries}
\title{Test function}
\usage{
timeseries(endpoint, id)
}
\arguments{
\item{endpoint}{An object of class \code{SensorwebEndpoint}}

\item{id}{The identifier of the timeseries to request from the endpoint}
}
\description{
Test function
}

