% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-tvp.R
\docType{class}
\name{TVP-class}
\alias{TVP}
\alias{TVP-class}
\alias{as.TVP}
\alias{is.TVP}
\title{Time-Value-Pairs}
\usage{
is.TVP(x)

as.TVP(x)

TVP(time = character(), value = numeric())
}
\description{
Class to contains the data of a \linkS4class{Timeseries} as time value
pairs.
}
\section{Functions}{
\itemize{
\item \code{is.TVP}: Checks whether \code{x} is a \code{TVP}.

\item \code{as.TVP}: Coerces \code{x} into a \code{TVP}.

\item \code{TVP}: Constructs a new \code{TVP}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

