% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-endpoint.R, R/class-timeseries.R, R/generic-methods.R, R/virtual-class-api-resource.R, R/virtual-class-http-resource.R
\docType{methods}
\name{resourceURL,Endpoint-method}
\alias{categoriesURL}
\alias{categoriesURL,Endpoint-method}
\alias{featuresURL}
\alias{featuresURL,Endpoint-method}
\alias{getDataURL}
\alias{getDataURL,Timeseries-method}
\alias{offeringsURL}
\alias{offeringsURL,Endpoint-method}
\alias{phenomenaURL}
\alias{phenomenaURL,Endpoint-method}
\alias{proceduresURL}
\alias{proceduresURL,Endpoint-method}
\alias{resourceURL}
\alias{resourceURL,ApiResource-method}
\alias{resourceURL,Endpoint-method}
\alias{servicesURL}
\alias{servicesURL,Endpoint-method}
\alias{stationsURL}
\alias{stationsURL,Endpoint-method}
\alias{subresourceURL}
\alias{subresourceURL,HttpResource-method}
\alias{timeseriesURL}
\alias{timeseriesURL,Endpoint-method}
\alias{url-methods}
\title{URL Methods}
\usage{
\S4method{resourceURL}{Endpoint}(x)

\S4method{stationsURL}{Endpoint}(x)

\S4method{servicesURL}{Endpoint}(x)

\S4method{timeseriesURL}{Endpoint}(x)

\S4method{categoriesURL}{Endpoint}(x)

\S4method{offeringsURL}{Endpoint}(x)

\S4method{featuresURL}{Endpoint}(x)

\S4method{proceduresURL}{Endpoint}(x)

\S4method{phenomenaURL}{Endpoint}(x)

\S4method{getDataURL}{Timeseries}(x)

resourceURL(x)

subresourceURL(x, ...)

stationsURL(x)

servicesURL(x)

timeseriesURL(x)

categoriesURL(x)

offeringsURL(x)

featuresURL(x)

proceduresURL(x)

phenomenaURL(x)

getDataURL(x)

\S4method{resourceURL}{ApiResource}(x)

\S4method{subresourceURL}{HttpResource}(x, ...)
}
\description{
Gets the URL of a resource.
}
\details{
\code{resourceURL} gets the URL of the resource \code{x}.

\code{subresourceURL} gets the URL of a sub-resource of \code{x}.

\code{stationsURL} gets the URL of a \linkS4class{Station} collection associated with \code{x}.

\code{servicesURL} gets the URL of a \linkS4class{Service} collection associated with \code{x}.

\code{timeseriesURL} gets the URL of a \linkS4class{Timeseries} collection associated with \code{x}.

\code{categoriesURL} gets the URL of a \linkS4class{Category} collection associated with \code{x}.

\code{offeringsURL} gets the URL of a \linkS4class{Offering} collection associated with \code{x}.

\code{featuresURL} gets the URL of a \linkS4class{Feature} collection associated with \code{x}.

\code{proceduresURL} gets the URL of a \linkS4class{Procedure} collection associated with \code{x}.

\code{phenomenaURL} gets the URL of a \linkS4class{Phenomenon} collection associated with \code{x}.

\code{phenomenaURL} gets the URL of the \code{getData} endpoint associated with \code{x}.
}
\author{
Christian Autermann \email{c.autermann@52north.org}
}

