% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-matrix.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\alias{read_distance_matrix}
\alias{save_distance_matrix}
\title{Distance Matrix Creation}
\usage{
create_distance_matrix(x)

save_distance_matrix(dm, file)

read_distance_matrix(file)
}
\arguments{
\item{x}{a Nx2 matrix of WGS84 coordinates or a \code{sp::SpatialPointsDataFrame}}

\item{dm}{the distance matrix}

\item{file}{the file name}

\item{file}{the file name}
}
\value{
the distance matrix of class \code{dist}
}
\description{
Creates a distance matrix for the supplied coordinates.

Saves a distance matrix to disk.

Reads the a distance matrix from disk.
}
\section{Functions}{
\itemize{
\item \code{save_distance_matrix}: 

\item \code{read_distance_matrix}: 
}}
\examples{
\dontrun{
stations <- get_stations(url)
dm <- create_distance_matrix(stations)
}
\dontrun{
save_distance_matrix(dm, 'distance_matrix.rds')
}
\dontrun{
dm <- read_distance_matrix('distance_matrix.rds')
}
}

