% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-endpoint.R
\name{example.endpoints}
\alias{example.endpoints}
\title{Example API endpoints.}
\usage{
example.endpoints()
}
\value{
R object with the further endpoints offered by the service
}
\description{
\code{sensorweb_api_endpoints} returns an instance of \linkS4class{Endpoint}
that can be used for testing.
}
\examples{
example.endpoints()
services(example.endpoints()[1])
}
\author{
Daniel Nuest \email{d.nuest@52north.org}

Christian Autermann \email{c.autermann@52north.org}
}

