% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sensorweb4R}
\alias{sensorweb4R}
\alias{sensorweb4R-package}
\title{\code{sensorweb4R}: functions and classes to download data from sensor web
services}
\description{
\code{sensorweb4R}: functions and classes to download data from sensor web
services
}
\section{About}{
 Sensor web services contain timeseries of sensoric data, such
  as temperature and rainfall by weather stations. This package can retrieve
  specific subsets of data, such as specific stations or observed phenomena
  by providin plain R function calls to load these datasets into your session
  as ready-to-use \code{R} data structures.
}

\section{Documentation}{
 The main documentation for this package can be
  found in the vignette (\code{vignette("sensorweb4R")}) and the demos
  (\code{demo(package = "sensorweb4R")}).
}

\section{Package options}{
 sensorweb4R uses the following
  \code{\link{options}} to configure behaviour:

  \itemize{ \item \code{sensorweb4R.loggerLayout}: layout string used in
  \code{futile.logger::flog.layout(..)} }
}

\section{Debugging}{
 sensorweb4R uses \code{futile.logger} for logging. To
  increase the log level of this package call \code{flog.threshold("<level>",
  name = "sensorweb4R")}. For more information about available logging levels
  or advanced logger configuration, see \code{?futile.logger}.
}
\author{
Christian Autermann, \email{c.autermann@52north.org}

Daniel Nuest, \email{d.nuest@52north.org}
}
\concept{
sensorweb, timeseries
}
\keyword{spatial}
\keyword{ts,}

