% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-timeseries.R
\docType{class}
\name{Timeseries-class}
\alias{Timeseries}
\alias{Timeseries-class}
\alias{as.Timeseries}
\alias{is.Timeseries}
\title{Timeseries}
\usage{
is.Timeseries(x)

as.Timeseries(x)

Timeseries(id = character(), label = NULL, uom = NULL, endpoint = NULL,
  phenomenon = NULL, service = NULL, feature = NULL, offering = NULL,
  procedure = NULL, category = NULL, station = NULL,
  statusIntervals = NULL, firstValue = NULL, lastValue = NULL,
  referenceValues = NULL)
}
\description{
Represents a timeseries.
}
\section{Functions}{
\itemize{
\item \code{is.Timeseries}: Checks whether \code{x} is a \code{Timeseries}.

\item \code{as.Timeseries}: Coerces \code{x} into a \code{Timeseries}.

\item \code{Timeseries}: Constructs a new \code{Timeseries}.
}}
\author{
Christian Autermann \email{c.autermann@52north.org}
}
\seealso{
Other API.Resources: \code{\link{ApiResource-class}},
  \code{\link{as.ApiResource}},
  \code{\link{as.list.ApiResource}}, \code{\link{fromURI}},
  \code{\link{is.ApiResource}},
  \code{\link{unique.ApiResource}}; \code{\link{Category}},
  \code{\link{Category-class}}, \code{\link{as.Category}},
  \code{\link{is.Category}}; \code{\link{Feature}},
  \code{\link{Feature-class}}, \code{\link{as.Feature}},
  \code{\link{is.Feature}}; \code{\link{Offering}},
  \code{\link{Offering-class}}, \code{\link{as.Offering}},
  \code{\link{is.Offering}}; \code{\link{Phenomenon}},
  \code{\link{Phenomenon-class}},
  \code{\link{as.Phenomenon}}, \code{\link{is.Phenomenon}};
  \code{\link{Procedure}}, \code{\link{Procedure-class}},
  \code{\link{as.Procedure}}, \code{\link{is.Procedure}};
  \code{\link{Service}}, \code{\link{Service-class}},
  \code{\link{as.Service}}, \code{\link{is.Service}};
  \code{\link{Station}}, \code{\link{Station-class}},
  \code{\link{as.Station}}, \code{\link{is.Station}}
}

